api/src/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
replicate>=0.22.0
mux-python>=3.15.0
ffmpeg-python>=0.2.0
requests>=2.31.0
python-multipart>=0.0.6  # For handling form data
pydantic>=2.6.0  # For data validation
pytest>=8.0.0  # For testing
httpx>=0.26.0  # Required for FastAPI TestClient
